refactor: 웹소켓 이벤트 스키마를 메세지 엔티티로 이동 (#94)
