+++ src/entities/message/model/socket-schema.ts
